refactoring: Sections 일급객체 생성
