+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -12,7 +12,6 @@ import java.util.*;
 @Entity
 public class Line extends BaseEntity {
 
-    private static final int SECTION_SIZE_LIMIT = 1;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -21,8 +20,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true, fetch = FetchType.EAGER)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -35,7 +34,7 @@ public class Line extends BaseEntity {
     public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
-        sections.add(Section.of(this, upStation, downStation, distance));
+        sections.addSection(Section.of(this, upStation, downStation, distance));
     }
 
     public void update(Line line) {
@@ -56,7 +55,7 @@ public class Line extends BaseEntity {
     }
 
     public List<Section> getSections() {
-        return sections;
+        return sections.getSections();
     }
 
     @Override
@@ -74,102 +73,14 @@ public class Line extends BaseEntity {
 
     public void addSection(Station upStation, Station downStation, int distance) {
         Section section = Section.of(this, upStation, downStation, distance);
-
-        if (isEmptyStations()) {
-            sections.add(section);
-            return;
-        }
-        checkAddSection(upStation, downStation);
-        sections.add(section);
-    }
-
-    private void checkAddSection(Station upStation, Station downStation) {
-        checkValidUpStation(upStation);
-        checkValidDownStation(downStation);
-    }
-
-    private boolean isEmptyStations() {
-        return getStations().size() == 0;
-    }
-
-    private void checkValidDownStation(Station downStation) {
-        boolean isDownStationExisted = getStations().contains(downStation);
-        if (isDownStationExisted) {
-            throw new AlreadyExistDownStation();
-        }
-    }
-
-    private void checkValidUpStation(Station upStation) {
-        boolean isNotValidUpStation = getLastStation() != upStation;
-        if (isNotValidUpStation) {
-            throw new InValidUpStationException();
-        }
-    }
-
-    private Station getLastStation() {
-        return getStations().get(getStations().size() - 1);
+        sections.addSection(section);
     }
 
     public List<Station> getStations() {
-        if (sections.isEmpty()) {
-            return Arrays.asList();
-        }
-
-        List<Station> stations = new ArrayList<>();
-        Station downStation = findUpStation();
-        stations.add(downStation);
-
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getUpStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getDownStation();
-            stations.add(downStation);
-        }
-
-        return stations;
-    }
-
-    private Station findUpStation() {
-        Station downStation = sections.get(0).getUpStation();
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getDownStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getUpStation();
-        }
-
-        return downStation;
+        return sections.getStations();
     }
 
     public void removeSection(Station station) {
-        checkSectionSize();
-        checkRemoveSection(station);
-
-        getSections().stream()
-                .filter(it -> it.getDownStation().equals(station))
-                .findFirst()
-                .ifPresent(it -> getSections().remove(it));
-    }
-
-    private void checkRemoveSection(Station station) {
-        boolean isNotValidUpStation = !getLastStation().equals(station);
-        if (isNotValidUpStation) {
-            throw new RuntimeException("하행 종점역만 삭제가 가능합니다.");
-        }
-    }
-
-    private void checkSectionSize() {
-        if (getSections().size() <= SECTION_SIZE_LIMIT) {
-            throw new InValidSectionSizeException();
-        }
+        sections.removeSection(station);
     }
 }
