+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -41,5 +41,4 @@ public class Station extends BaseEntity {
         return Objects.hash(id, name);
     }
 
-
 }
